requirements.txt
Flask
python-dotenv
spacy
trafilatura
newspaper3k
fpdf
matplotlib
wordcloud
textblob
groq
